fix : cars 일급 컬렉션을 통해 자동차 갯수를 확인이 가능하기때문에 불필요 객체 삭제
